Boost version to 0.9.0 * Add property for virtual pin * Improve GSM ESP32 check size * Add ota ESP32, ESP8266, RPi * Add control via BLE ESP32 * Add multi client * Add keyworks * Fix ESP8266 version 3.1.1 ...

+++ src/Modbus/ERaModbus.hpp
@@ -1,8 +1,7 @@
 #ifndef INC_ERA_MODBUS_HPP_
 #define INC_ERA_MODBUS_HPP_
 
-#include <vector>
-#include <cmath>
+#include <math.h>
 #include <ERa/ERaParam.hpp>
 #include <Utility/ERaUtility.hpp>
 #include <Modbus/ERaModbusState.hpp>
@@ -29,6 +28,7 @@ class ERaModbus
     const char* TAG = "Modbus";
     const char* FILENAME_CONFIG = FILENAME_MODBUS_CONFIG;
     const char* FILENAME_CONTROL = FILENAME_MODBUS_CONTROL;
+    const IPAddress ipNone{0, 0, 0, 0};
 
     friend class ERaModbusTransp < ERaModbus<Api> >;
 	typedef ERaModbusTransp < ERaModbus<Api> > ModbusTransp;
@@ -37,12 +37,15 @@ public:
     ERaModbus()
         : modbusConfig(ERaApplication::config())
         , modbusControl(ERaApplication::control())
+        , timeout(DEFAULT_TIMEOUT_MODBUS)
+        , prevMillis(0)
         , total(0)
         , failRead(0)
         , failWrite(0)
         , dePin(-1)
         , transp(0)
-        , client(NULL)
+        , streamRTU(NULL)
+        , clientTCP(NULL)
         , ip {
             .ip = {},
             .port = 502
@@ -59,23 +62,25 @@ public:
         this->switchToModbusRTU();
     }
 
-    void setModbusClient(Client& _client, IPAddress _ip, uint16_t _port = 502) {
-        this->client = &_client;
+    void setModbusClient(Client& _client, IPAddress _ip = IPAddress(0, 0, 0, 0), uint16_t _port = 502) {
+        this->clientTCP = &_client;
         this->ip.ip = _ip;
         this->ip.port = _port;
-        this->setModbusStream(_client);
     }
 
     void setModbusStream(Stream& _stream) {
-        this->stream = &_stream;
+        this->streamRTU = &_stream;
     }
 
-    void switchToModbusRTU() {
-        this->transp = ModbusTransportT::MODBUS_TRANSPORT_RTU;
+    void setModbusTimeout(uint32_t _timeout) {
+        this->timeout = _timeout;
     }
 
-    void switchToModbusTCP() {
-        this->transp = ModbusTransportT::MODBUS_TRANSPORT_TCP;
+    void setPubModbusInterval(uint32_t _interval) {
+        if (!_interval) {
+            return;
+        }
+        ERaApplication::getConfig()->pubInterval.delay = _interval;
     }
 
     void setModbusDEPin(int _pin);
@@ -91,6 +96,7 @@ protected:
     }
 
     void run(const bool isRead) {
+        this->waitTCPIpReady();
         for (;;) {
             if (isRead) {
                 this->runRead();
@@ -126,22 +132,35 @@ protected:
         req.key = nullptr;
     }
 
-    void parseModbusConfig(char* ptr, bool isControl = false) {
+    void parseModbusConfig(char* config, const char* hash, const char* buf, bool isControl = false) {
         if (isControl) {
-            this->modbusControl->parseConfig(ptr);
-            this->thisApi().writeToFlash(FILENAME_CONTROL, ptr);
+            ERaGuardLock(this->mutex);
+            this->modbusControl->parseConfig(config);
+            if (this->modbusControl->updateHashID(hash)) {
+                this->thisApi().writeToFlash(FILENAME_CONTROL, buf);
+            }
+            ERaGuardUnlock(this->mutex);
         } else {
-            this->modbusConfig->parseConfig(ptr);
+            ERaGuardLock(this->mutex);
+            this->modbusConfig->parseConfig(config);
+            if (this->modbusConfig->updateHashID(hash)) {
+                this->clearDataBuff();
+                this->thisApi().writeToFlash(FILENAME_CONFIG, buf);
+            }
+            ERaGuardUnlock(this->mutex);
             this->setBaudRate(this->modbusConfig->baudSpeed);
             this->executeNow();
-            this->thisApi().writeToFlash(FILENAME_CONFIG, ptr);
         }
         ModbusState::set(ModbusStateT::STATE_MB_PARSE);
     }
 
     void removeConfigFromFlash() {
-        this->modbusConfig->deleteAll();
-        this->modbusControl->deleteAll();
+        if (this->modbusConfig != nullptr) {
+            this->modbusConfig->deleteAll();
+        }
+        if (this->modbusControl != nullptr) {
+            this->modbusControl->deleteAll();
+        }
         this->thisApi().removeFromFlash(FILENAME_CONFIG);
         this->thisApi().removeFromFlash(FILENAME_CONTROL);
     }
@@ -157,8 +176,12 @@ protected:
 	    if (!this->queue.writeable()) {
             return false;
         }
+        char* buf = (char*)ERA_MALLOC(37 * sizeof(char));
+        if (buf == nullptr) {
+            return false;
+        }
         ModbusAction_t req;
-        req.key = (char*)ERA_MALLOC(37 * sizeof(char));
+        req.key = buf;
         memset(req.key, 0, 37 * sizeof(char));
         strcpy(req.key, ptr);
         this->queue += req;
@@ -179,28 +202,75 @@ private:
     void initModbusConfig() {
         char* ptr = nullptr;
         ptr = this->thisApi().readFromFlash(FILENAME_CONFIG);
-        this->modbusConfig->parseConfig(ptr);
+        this->updateConfig(this->modbusConfig, ptr, "configuration");
         this->setBaudRate(this->modbusConfig->baudSpeed);
         free(ptr);
         ptr = this->thisApi().readFromFlash(FILENAME_CONTROL);
-        this->modbusControl->parseConfig(ptr);
+        this->updateConfig(this->modbusControl, ptr, "control");
         free(ptr);
         ptr = nullptr;
     }
 
-    bool connectTCPIp() {
-        if (this->client == nullptr) {
-            return true;
+    void updateConfig(ERaApplication* config, const char* buf, const char* name) {
+        cJSON* root = cJSON_Parse(buf);
+        if (!cJSON_IsObject(root)) {
+            cJSON_Delete(root);
+            root = nullptr;
+            return;
         }
-        if (!this->client->connected()) {
-            return this->client->connect(ip.ip, ip.port);
+
+        cJSON* data = cJSON_GetObjectItem(root, "data");
+        cJSON* item = cJSON_GetObjectItem(data, "hash_id");
+        if (cJSON_IsString(item)) {
+            config->updateHashID(item->valuestring);
         }
-        return true;
+        item = cJSON_GetObjectItem(data, name);
+        if (cJSON_IsString(item)) {
+            config->parseConfig(item->valuestring);
+        }
+
+        cJSON_Delete(root);
+        root = nullptr;
+    }
+
+    void nextTransport(const ModbusConfig_t& param) {
+        this->switchToModbusRTU();
+        if (this->clientTCP == nullptr) {
+            return;
+        }
+        if (param.ipSlave.ip.dword) {
+            this->switchToModbusTCP();
+        }
+        if ((ip.ip == IPAddress(param.ipSlave.ip.dword)) ||
+            (param.ipSlave.ip.dword == 0)) {
+            return;
+        }
+        if (param.ipSlave.port == 0) {
+            return;
+        }
+        ip.ip = param.ipSlave.ip.dword;
+        ip.port = param.ipSlave.port;
+        // TODO
+        if (this->clientTCP->connected()) {
+            this->clientTCP->stop();
+        }
+        this->clientTCP->connect(ip.ip, ip.port);
+    }
+
+    void switchToModbusRTU() {
+        this->stream = this->streamRTU;
+        this->transp = ModbusTransportT::MODBUS_TRANSPORT_RTU;
+    }
+
+    void switchToModbusTCP() {
+        this->stream = this->clientTCP;
+        this->transp = ModbusTransportT::MODBUS_TRANSPORT_TCP;
     }
 
     void configModbus();
     void setBaudRate(uint32_t baudrate);
     void readModbusConfig();
+    bool checkPubDataInterval();
     void delayModbus(const int address);
     ModbusConfigAlias_t* getModbusAlias(const char* key);
     ModbusConfig_t* getModbusConfig(int id);
@@ -214,6 +284,15 @@ private:
     void sendCommand(uint8_t* data, size_t size);
     void switchToTransmit();
     void switchToReceive();
+
+    void waitTCPIpReady() {
+        if (this->clientTCP == nullptr) {
+            return;
+        }
+        do {
+            ERaDelay(1000);
+        } while (!this->thisApi().connected());
+    }
 
     bool getBit(uint8_t byte, uint8_t pos) {
         return ((byte >> pos) & 0x01);
@@ -253,12 +332,15 @@ private:
     ERaDataBuffDynamic dataBuff;
     ERaApplication*& modbusConfig;
     ERaApplication*& modbusControl;
+    uint32_t timeout;
+    unsigned long prevMillis;
     int total;
     int failRead;
     int failWrite;
     int dePin;
     uint8_t transp;
-    Client* client;
+    Stream* streamRTU;
+    Client* clientTCP;
     TCPIp_t ip;
     Stream* stream;
     bool _streamDefault;
@@ -273,33 +355,58 @@ void ERaModbus<Api>::readModbusConfig() {
     if (this->modbusConfig->modbusConfigParam.isEmpty()) {
         return;
     }
-    if (!this->connectTCPIp()) {
-        return;
-    }
     this->dataBuff.clear();
-    for (size_t i = 0; i < this->modbusConfig->readConfigCount; ++i) {
-        ModbusConfig_t& param = *this->modbusConfig->modbusConfigParam[i];
+    const ERaList<ModbusConfig_t*>::iterator* e = this->modbusConfig->modbusConfigParam.end();
+    for (ERaList<ModbusConfig_t*>::iterator* it = this->modbusConfig->modbusConfigParam.begin(); it != e; it = it->getNext()) {
+        ModbusConfig_t* param = it->get();
+        if (param == nullptr) {
+            continue;
+        }
         ERaGuardLock(this->mutex);
-        this->sendModbusRead(param);
+        this->sendModbusRead(*param);
         ERaGuardUnlock(this->mutex);
-        this->delayModbus(param.addr);
+        this->delayModbus(param->addr);
         if (ModbusState::is(ModbusStateT::STATE_MB_PARSE) ||
             ModbusState::is(ModbusStateT::STATE_MB_CONTROLLED)) {
             this->executeNow();
             return;
         }
     }
+    if (this->dataBuff.isEmpty()) {
+        return;
+    }
+    if (this->dataBuff.isChange()) {
+        this->prevMillis = ERaMillis();
+    }
+    else if (!this->checkPubDataInterval()) {
+        return;
+    }
     this->dataBuff.add_multi(ERA_F("fail_read"), this->failRead, ERA_F("fail_write"), this->failWrite, ERA_F("total"), this->total);
     this->thisApi().modbusDataWrite(&this->dataBuff);
+}
+
+template <class Api>
+bool ERaModbus<Api>::checkPubDataInterval() {
+    unsigned long currentMillis = ERaMillis();
+    if (currentMillis - this->prevMillis < this->modbusConfig->pubInterval.delay) {
+        return false;
+    }
+    unsigned long skipTimes = (currentMillis - this->prevMillis) / this->modbusConfig->pubInterval.delay;
+    this->prevMillis += this->modbusConfig->pubInterval.delay * skipTimes;
+    return true;
 }
 
 template <class Api>
 void ERaModbus<Api>::delayModbus(const int address) {
     MillisTime_t delayMs {10L};
     MillisTime_t startMillis = ERaMillis();
-    for (ERaList<SensorDelay_t>::iterator* it = this->modbusConfig->sensorDelay.begin(); it != nullptr; it = it->getNext()) {
-        if (it->get().address == address) {
-            delayMs = it->get().delay;
+    const ERaList<SensorDelay_t*>::iterator* e = this->modbusConfig->sensorDelay.end();
+    for (ERaList<SensorDelay_t*>::iterator* it = this->modbusConfig->sensorDelay.begin(); it != e; it = it->getNext()) {
+        if (it->get() == nullptr) {
+            continue;
+        }
+        if (it->get()->address == address) {
+            delayMs = it->get()->delay;
             break;
         }
     }
@@ -318,9 +425,13 @@ void ERaModbus<Api>::delayModbus(const int address) {
 
 template <class Api>
 ModbusConfigAlias_t* ERaModbus<Api>::getModbusAlias(const char* key) {
-    for (ERaList<ModbusConfigAlias_t>::iterator* it = this->modbusConfig->modbusConfigAliasParam.begin(); it != nullptr; it = it->getNext()) {
-        if (!strcmp(it->get().key, key)) {
-            return &it->get();
+    const ERaList<ModbusConfigAlias_t*>::iterator* e = this->modbusConfig->modbusConfigAliasParam.end();
+    for (ERaList<ModbusConfigAlias_t*>::iterator* it = this->modbusConfig->modbusConfigAliasParam.begin(); it != e; it = it->getNext()) {
+        if (it->get() == nullptr) {
+            continue;
+        }
+        if (!strcmp(it->get()->key, key)) {
+            return it->get();
         }
     }
     return nullptr;
@@ -340,10 +451,6 @@ bool ERaModbus<Api>::actionModbus(const char* key) {
         return false;
     }
 
-    if (!this->connectTCPIp()) {
-        return false;
-    }
-
     ModbusConfig_t* config = nullptr;
     for (int i = 0; i < alias->readActionCount; ++i) {
         ERaGuardLock(this->mutex);
@@ -360,9 +467,13 @@ bool ERaModbus<Api>::actionModbus(const char* key) {
 
 template <class Api>
 ModbusConfig_t* ERaModbus<Api>::getModbusConfig(int id) {
-    for (ERaList<ModbusConfig_t>::iterator* it = this->modbusControl->modbusConfigParam.begin(); it != nullptr; it = it->getNext()) {
-        if (it->get().id == id) {
-            return &it->get();
+    const ERaList<ModbusConfig_t*>::iterator* e = this->modbusControl->modbusConfigParam.end();
+    for (ERaList<ModbusConfig_t*>::iterator* it = this->modbusControl->modbusConfigParam.begin(); it != e; it = it->getNext()) {
+        if (it->get() == nullptr) {
+            continue;
+        }
+        if (it->get()->id == id) {
+            return it->get();
         }
     }
     return nullptr;
@@ -386,6 +497,7 @@ bool ERaModbus<Api>::eachActionModbus(Action_t& action, ModbusConfig_t*& config)
 template <class Api>
 void ERaModbus<Api>::sendModbusRead(ModbusConfig_t& param) {
     bool status {false};
+    this->nextTransport(param);
     switch (param.func) {
         case ModbusFunctionT::READ_COIL_STATUS:
             status = ModbusTransp::readCoilStatus(this->transp, param);
@@ -415,6 +527,7 @@ void ERaModbus<Api>::sendModbusRead(ModbusConfig_t& param) {
 template <class Api>
 bool ERaModbus<Api>::sendModbusWrite(ModbusConfig_t& param) {
     bool status {false};
+    this->nextTransport(param);
     switch (param.func) {
         case ModbusFunctionT::FORCE_SINGLE_COIL:
             status = ModbusTransp::forceSingleCoil(this->transp, param);
@@ -448,15 +561,16 @@ void ERaModbus<Api>::onData(ERaModbusRequest* request, ERaModbusResponse* respon
     switch (request->getFunction()) {
         case ModbusFunctionT::READ_COIL_STATUS:
         case ModbusFunctionT::READ_INPUT_STATUS: {
+            LOC_BUFFER_MODBUS(1)
             uint16_t bits = request->getLength();
-            uint8_t pData[request->getLength()] {0};
             for (int i = 0; (i < response->getBytes()) && (bits > 0); ++i) {
                 for (int j = 0; j < ((bits > 8) ? 8 : bits); ++j) {
                     pData[i * 8 + j] = this->getBit(response->getData()[i], j);
                 }
                 bits -= ((bits > 8) ? 8 : bits);
             }
-            this->dataBuff.add_hex_array(pData, request->getLength());
+            this->dataBuff.add_hex_array(pData, pDataLen);
+            FREE_BUFFER_MODBUS
         }
             break;
         default:
@@ -479,14 +593,14 @@ void ERaModbus<Api>::onError(ERaModbusRequest* request) {
         case ModbusFunctionT::READ_COIL_STATUS:
         case ModbusFunctionT::READ_INPUT_STATUS:
             if (!this->dataBuff.next()) {
-                uint8_t pData[request->getLength()] {0};
-                this->dataBuff.add_hex_array(pData, request->getLength());
+                size_t pDataLen = request->getLength();
+                this->dataBuff.add_zero_array(pDataLen);
             }
             break;
         default:
             if (!this->dataBuff.next()) {
-                uint8_t pData[request->getLength() * 2] {0};
-                this->dataBuff.add_hex_array(pData, request->getLength() * 2);
+                size_t pDataLen = request->getLength() * 2;
+                this->dataBuff.add_zero_array(pDataLen);
             }
             break;
     }
@@ -498,6 +612,7 @@ void ERaModbus<Api>::setModbusDEPin(int _pin) {
     if (_pin < 0) {
         return;
     }
+
     this->dePin = _pin;
 #if defined(ARDUINO) || \
     (defined(LINUX) &&  \
@@ -511,6 +626,7 @@ void ERaModbus<Api>::switchToTransmit() {
     if (this->dePin < 0) {
         return;
     }
+
 #if defined(ARDUINO) || \
     (defined(LINUX) &&  \
     defined(RASPBERRY))
@@ -523,6 +639,7 @@ void ERaModbus<Api>::switchToReceive() {
     if (this->dePin < 0) {
         return;
     }
+
 #if defined(ARDUINO) || \
     (defined(LINUX) &&  \
     defined(RASPBERRY))
